Moved Layer and ILayer to new Layer package
