+++ src/main/java/C7/Layer/Layer.java
@@ -1,4 +1,6 @@
-package C7;
+package C7.Layer;
+
+import C7.Color;
 
 /**
  * Represents a basic image layer.
